Fix semantics pblm for HTML/Accessibility, import logements.json, continue homepage
