+++ src/components/card/card.css
@@ -1,0 +1,36 @@
+.cardWrapper {
+    position: relative;
+    display: flex;
+    border-radius: 10px;
+    background-image: linear-gradient(to bottom, #FD6868 0%, rgba(0, 0, 0, 0.9) 90%);
+}
+
+.figureCard {
+    position: relative;
+    overflow: hidden;
+    border-radius: 10px;
+}
+
+.imageCard {
+    object-fit: cover;
+}
+
+.cardWrapper:hover {
+    cursor: pointer;
+}
+
+.figcaptionCard {
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    background-image: linear-gradient(to bottom, transparent, rgba(0, 0, 0, 0.25), rgba(0, 0, 0, 0.7));
+}
+
+.cardTitle {
+    color: white;
+    font-family: 'Montserrat', serif;
+    font-weight: 400;
+    font-size: 1rem;
+    padding: 20px 235px 15px 20px
+}
