+++ src/pages/home/home.css
@@ -1,0 +1,35 @@
+.divContainer {
+    position: relative;
+    height: 225px;
+}
+
+.imgContainer {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 25px;
+    filter: brightness(0.8);
+}
+
+.h1Container {
+    font-family: 'Montserrat', serif;
+    color: white;
+    font-weight: 400;
+    font-size: 3rem;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+}
+
+.cardContainer {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    justify-items: center;
+    align-items: center;
+    margin-top: 45px;
+    width: 100%;
+    height: 830px;
+    border-radius: 25px;
+    background-color: #f6f6f6;
+}
