--- src/utils/styles/fonts.js
@@ -1,13 +1,0 @@
-const fonts = {
-
-    primary: {
-        name: 'Montserrat',
-        weight: '400',
-        style: 'normal',
-        size: '16px',
-        link: 'https://fonts.googleapis.com/css2?family=Montserrat:wght@400&display=swap'
-    }
-
-}
-
-export default fonts
